refactor: move each test group into separate files
