+++ tests/unit/tests/Framework/QueryBuilder/OrderByTest.php
@@ -1,0 +1,41 @@
+<?php
+
+use Give\Framework\Database\DB;
+use Give\Framework\QueryBuilder\QueryBuilder;
+use PHPUnit\Framework\TestCase;
+
+final class OrderByTest extends TestCase
+{
+    public function testOrderByDesc()
+    {
+        $builder = new QueryBuilder();
+
+        $builder
+            ->select('*')
+            ->from(DB::raw('posts'))
+            ->where('post_status', 'published')
+            ->orderBy('ID', 'DESC');
+
+        $this->assertContains(
+            "SELECT * FROM posts WHERE post_status = 'published' ORDER BY ID DESC",
+            $builder->getSQL()
+        );
+    }
+
+
+    public function testOrderByAsc()
+    {
+        $builder = new QueryBuilder();
+
+        $builder
+            ->select('*')
+            ->from(DB::raw('posts'))
+            ->where('post_status', 'published')
+            ->orderBy('ID', 'ASC');
+
+        $this->assertContains(
+            "SELECT * FROM posts WHERE post_status = 'published' ORDER BY ID ASC",
+            $builder->getSQL()
+        );
+    }
+}
